+++ src/main/java/com/bychenkv/controller/ExchangeRatesServlet.java
@@ -8,6 +8,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.util.List;
 
 @WebServlet("/exchangeRates")
@@ -30,7 +31,7 @@ public class ExchangeRatesServlet extends BaseServlet {
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException {
         String base = getCurrencyCodeParameter(req, "baseCurrencyCode");
         String target = getCurrencyCodeParameter(req, "targetCurrencyCode");
-        double rate = getRateParameter(req);
+        BigDecimal rate = getRateParameter(req);
 
         ExchangeRate exchangeRate = dao.save(new CurrencyCodePair(base, target), rate);
         sendJson(resp, HttpServletResponse.SC_CREATED, exchangeRate);
